+++ CHANGELOG.md
@@ -1,13 +1,28 @@
 # 0.8.0
 
-This update finalizes End Cities.
+This update finalizes End Villages.
 
 ## Entities
 **[*] Pixie**
 - Has a much contained spawn area, preventing pixies spawning underground
 
+**[*] End Villagers**
+- All trades have been added, with placeholders for unimplemented items.
+- [Overview of all changed trades](https://docs.google.com/spreadsheets/d/1BkHnG_Mgk8K887Ls4g4pnAhQ7jwYTwwm9QV21W9t9uM/edit?gid=1755175231#gid=1755175231)
+
 ## Items
 **[+] Loaf of Plenty**
+
+**[+] Candied Chorus Fruit**
+
+**[*] Fried Chorus Fruit**
+- Has a cooldown of 1s now
+
+**[*] Golden Chorus Fruit**
+- Has a cooldown of 1s now
+
+**[*] Shulker Body**
+- Has a cooldown of 1s now
 
 ## Mechanics
 
@@ -17,4 +32,4 @@ This update finalizes End Cities.
 
 # Developer's Note
 
-End cities go brrrrrrrrrrrrrrrrr
+End villages go brrrrrrrrrrrrrrrrr